Added Nav Toggle
Temporary Nav toggle while waiting for Sams implementation.
Cursor turns to hand while hovering over nav h2s.
Header siblings are hidden when header has a class of hide.
Header’s have a class of hide toggled when clicked.
When the nav is loaded in php, the current module is shown and the rest
of the nav pages are hidden, if there is no current module then the
first module in the nav is shown and the rest are hidden.
This will potentially be replaced by a similar jQuery script that also
animates the change.